first draft of a simple python based virus
this simple virus will search the folder (the file is in) and all subfolders for .py files. It will then try to infect these with its code and signature. Outbreak (bomb) has not been define.

This is just for academic purpose and it is relatively harmless. But be warned that it does modify .py files it finds and might render them useless. It is a virus after all.